Fix DeleteActivity: only activity's owner can delete the activity

+++ activity/service.go
@@ -17,7 +17,7 @@ type Service interface {
 	GetByID(activityID *uint) (*ActivityResponse, int, error)
 	GetAll(userID *uint) (*[]ActivityResponse, int, error)
 	UploadActivityPhotoToS3(file *multipart.File, fileHeader *multipart.FileHeader) (*string, error)
-	DeleteActivity(request *DeleteActivityRequest) (int, error)
+	DeleteActivity(userID uint, request *DeleteActivityRequest) (int, error)
 
 	// Like
 	CreateLike(userID uint, like *LikeRequest) (int, error)
@@ -200,6 +200,9 @@ func (s *service) DeleteComment(userID, commentID uint) (int, error) {
 	if userID == commentOwner || userID == activityOwner {
 		err = s.repository.DeleteComment(commentID)
 		if err != nil {
+			if errors.Is(err, gorm.ErrRecordNotFound) {
+				return http.StatusNotFound, errors.New(constants.CommentNotFound)
+			}
 			return http.StatusInternalServerError, err
 		}
 		return http.StatusOK, nil
@@ -222,7 +225,16 @@ func (s *service) UploadActivityPhotoToS3(file *multipart.File, fileHeader *mult
 	return &avatarUrl, nil
 }
 
-func (s *service) DeleteActivity(request *DeleteActivityRequest) (int, error) {
+func (s *service) DeleteActivity(userID uint, request *DeleteActivityRequest) (int, error) {
+	activity, statusCode, err := s.GetByID(&request.ActivityID)
+	if err != nil {
+		return statusCode, err
+	}
+
+	if activity.User.ID != userID {
+		return http.StatusForbidden, errors.New(constants.ActivityDeleteAccessDenied)
+	}
+
 	if err := s.repository.DeleteActivityByID(request.ActivityID); err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
 			return http.StatusNotFound, errors.New(constants.ActivityNotFound)
